requirements.txt
[packages]
django = "==5.2.1"
pillow = ">=10.0.0"
django-crispy-forms = "==1.12.0"
django-allauth = "==0.45.0"
django-countries = "==7.2.1"
stripe = "==2.60.0"
python-dotenv = "==0.19.0"
psycopg = "==3.1.8"
whitenoise = "==5.3.0"
gunicorn = "==20.1.0"
djangorestframework = "*"
djoser = "*"
django-cors-headers = "*"
dj-rest-auth = "*"
sslcommerz_sdk = "*"

[dev-packages]

[requires]
python_version = "3.13"
